入力出力データを追加 & Kalman::Camera の中身を double に変更

+++ trashbox/header/Encoder.hh
@@ -15,7 +15,7 @@ namespace Kalman
     }Scale;
 
     typedef struct _camera{
-        int x,y;
+        double x,y;
     }Camera;
 
     class Encoder:dynamical_system
